+++ platform/darwin/src/MGLLineStyleLayer.h
@@ -229,6 +229,21 @@ MGL_EXPORT
  */
 @property (nonatomic, null_resettable) NSExpression *lineRoundLimit;
 
+/**
+ Sorts features in ascending order based on this value. Features with a higher
+ sort key will appear above features with a lower sort key.
+ 
+ You can set this property to an expression containing any of the following:
+ 
+ * Constant numeric values
+ * Predefined functions, including mathematical and string operators
+ * Conditional expressions
+ * Variable assignments and references to assigned variables
+ * Interpolation and step functions applied to the `$zoomLevel` variable and/or
+ feature attributes
+ */
+@property (nonatomic, null_resettable) NSExpression *lineSortKey;
+
 #pragma mark - Accessing the Paint Attributes
 
 /**