+++ platform/darwin/test/MGLLineStyleLayerTests.mm
@@ -219,6 +219,75 @@
         XCTAssertThrowsSpecificNamed(layer.lineRoundLimit = functionExpression, NSException, NSInvalidArgumentException, @"MGLLineLayer should raise an exception if a camera-data expression is applied to a property that does not support key paths to feature attributes.");
     }
 
+    // line-sort-key
+    {
+        XCTAssertTrue(rawLayer->getLineSortKey().isUndefined(),
+                      @"line-sort-key should be unset initially.");
+        NSExpression *defaultExpression = layer.lineSortKey;
+
+        NSExpression *constantExpression = [NSExpression expressionWithFormat:@"1"];
+        layer.lineSortKey = constantExpression;
+        mbgl::style::PropertyValue<float> propertyValue = { 1.0 };
+        XCTAssertEqual(rawLayer->getLineSortKey(), propertyValue,
+                       @"Setting lineSortKey to a constant value expression should update line-sort-key.");
+        XCTAssertEqualObjects(layer.lineSortKey, constantExpression,
+                              @"lineSortKey should round-trip constant value expressions.");
+
+        constantExpression = [NSExpression expressionWithFormat:@"1"];
+        NSExpression *functionExpression = [NSExpression expressionWithFormat:@"mgl_step:from:stops:($zoomLevel, %@, %@)", constantExpression, @{@18: constantExpression}];
+        layer.lineSortKey = functionExpression;
+
+        {
+            using namespace mbgl::style::expression::dsl;
+            propertyValue = mbgl::style::PropertyExpression<float>(
+                step(zoom(), literal(1.0), 18.0, literal(1.0))
+            );
+        }
+
+        XCTAssertEqual(rawLayer->getLineSortKey(), propertyValue,
+                       @"Setting lineSortKey to a camera expression should update line-sort-key.");
+        XCTAssertEqualObjects(layer.lineSortKey, functionExpression,
+                              @"lineSortKey should round-trip camera expressions.");
+
+        functionExpression = [NSExpression expressionWithFormat:@"mgl_interpolate:withCurveType:parameters:stops:(keyName, 'linear', nil, %@)", @{@18: constantExpression}];
+        layer.lineSortKey = functionExpression;
+
+        {
+            using namespace mbgl::style::expression::dsl;
+            propertyValue = mbgl::style::PropertyExpression<float>(
+                interpolate(linear(), number(get("keyName")), 18.0, literal(1.0))
+            );
+        }
+
+        XCTAssertEqual(rawLayer->getLineSortKey(), propertyValue,
+                       @"Setting lineSortKey to a data expression should update line-sort-key.");
+        NSExpression *pedanticFunctionExpression = [NSExpression expressionWithFormat:@"mgl_interpolate:withCurveType:parameters:stops:(CAST(keyName, 'NSNumber'), 'linear', nil, %@)", @{@18: constantExpression}];
+        XCTAssertEqualObjects(layer.lineSortKey, pedanticFunctionExpression,
+                              @"lineSortKey should round-trip data expressions.");
+
+        functionExpression = [NSExpression expressionWithFormat:@"mgl_interpolate:withCurveType:parameters:stops:($zoomLevel, 'linear', nil, %@)", @{@10: functionExpression}];
+        layer.lineSortKey = functionExpression;
+
+        {
+            using namespace mbgl::style::expression::dsl;
+            propertyValue = mbgl::style::PropertyExpression<float>(
+                interpolate(linear(), zoom(), 10.0, interpolate(linear(), number(get("keyName")), 18.0, literal(1.0)))
+            );
+        }
+
+        XCTAssertEqual(rawLayer->getLineSortKey(), propertyValue,
+                       @"Setting lineSortKey to a camera-data expression should update line-sort-key.");
+        pedanticFunctionExpression = [NSExpression expressionWithFormat:@"mgl_interpolate:withCurveType:parameters:stops:($zoomLevel, 'linear', nil, %@)", @{@10: pedanticFunctionExpression}];
+        XCTAssertEqualObjects(layer.lineSortKey, pedanticFunctionExpression,
+                              @"lineSortKey should round-trip camera-data expressions.");
+
+        layer.lineSortKey = nil;
+        XCTAssertTrue(rawLayer->getLineSortKey().isUndefined(),
+                      @"Unsetting lineSortKey should return line-sort-key to the default value.");
+        XCTAssertEqualObjects(layer.lineSortKey, defaultExpression,
+                              @"lineSortKey should return the default value after being unset.");
+    }
+
     // line-blur
     {
         XCTAssertTrue(rawLayer->getLineBlur().isUndefined(),
@@ -878,6 +947,7 @@
     [self testPropertyName:@"line-join" isBoolean:NO];
     [self testPropertyName:@"line-miter-limit" isBoolean:NO];
     [self testPropertyName:@"line-round-limit" isBoolean:NO];
+    [self testPropertyName:@"line-sort-key" isBoolean:NO];
     [self testPropertyName:@"line-blur" isBoolean:NO];
     [self testPropertyName:@"line-color" isBoolean:NO];
     [self testPropertyName:@"line-dash-pattern" isBoolean:NO];